Refactored logic and functions
Fully reimagined functions for control panel. Now its looking much better and comact, in addition its easy to read and maintain or add new features. Also its working much faster as it has less unnecessary code loops as it uses already saved data. In total much more perfoment action for the App, hovewer it's not possible to move the blocks on the canvas. Going to be fixed in next update